service/mockup/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0

